fixed name scaling
