+++ src/constants.ts
@@ -1,11 +1,3 @@
-export const cardImgWidth = 747;
-export const cardImgHeight = 1038;
-export const cardImgAspect = cardImgHeight / cardImgWidth;
-export const baseFontSize = 16;
-export const cardId = 'card';
-// Form width minus all paddings
-export const cropperWidth = 354;
-export const cropperHeight = (cropperWidth / cardImgWidth) * cardImgHeight;
 // 5 megabytes
 export const maxFileSize = 5_242_880;
