+++ src/features/cardEditor/cardOptions/components/blocks/ImagesForm/constants.ts
@@ -1,0 +1,5 @@
+import { cardImgHeight, cardImgWidth } from '@cardEditor/cardStyles';
+
+// Form width minus all paddings
+export const cropperWidth = 354;
+export const cropperHeight = (cropperWidth / cardImgWidth) * cardImgHeight;
